fix levels for fine faders

+++ src/state.h
@@ -137,7 +137,7 @@ typedef struct {
     u8 group;
     u8 x, y;
     u8 w, h;
-    s8 level;
+    s16 level;
     s8 script;
 } grid_common_t;
